+++ src/rust/madara/sources/pojokmanga/res/filters.json
@@ -1,0 +1,206 @@
+[
+	{
+		"type": "title"
+	},
+	{
+		"type": "group",
+		"name": "Status",
+		"filters": [
+			{
+				"type": "check",
+				"name": "Ongoing"
+			},
+			{
+				"type": "check",
+				"name": "On Hold"
+			},
+			{
+				"type": "check",
+				"name": "Cancelled"
+			},
+			{
+				"type": "check",
+				"name": "Completed"
+			}
+		]
+	},
+	{
+		"type": "select",
+		"name": "Genre Condition",
+		"options": [
+			"OR (having one of the selected genre)",
+			"AND (having all of the selected genres)"
+		]
+	},
+	{
+		"type": "group",
+		"name": "Genres",
+		"filters": [
+			{
+				"type": "genre",
+				"name": "Action",
+				"id": "action"
+			},
+			{
+				"type": "genre",
+				"name": "Adventure",
+				"id": "adventure"
+			},
+			{
+				"type": "genre",
+				"name": "Comedy",
+				"id": "comedy"
+			},
+			{
+				"type": "genre",
+				"name": "Cooking",
+				"id": "cooking"
+			},
+			{
+				"type": "genre",
+				"name": "Drama",
+				"id": "drama"
+			},
+			{
+				"type": "genre",
+				"name": "Fantasy",
+				"id": "fantasy"
+			},
+			{
+				"type": "genre",
+				"name": "Game",
+				"id": "game"
+			},
+			{
+				"type": "genre",
+				"name": "Gender Bender",
+				"id": "gender-bender"
+			},
+			{
+				"type": "genre",
+				"name": "Harem",
+				"id": "harem"
+			},
+			{
+				"type": "genre",
+				"name": "Historical",
+				"id": "historical"
+			},
+			{
+				"type": "genre",
+				"name": "Horror",
+				"id": "horror"
+			},
+			{
+				"type": "genre",
+				"name": "Isekai",
+				"id": "isekai"
+			},
+			{
+				"type": "genre",
+				"name": "Magic",
+				"id": "magic"
+			},
+			{
+				"type": "genre",
+				"name": "Manga",
+				"id": "manga"
+			},
+			{
+				"type": "genre",
+				"name": "Manhua",
+				"id": "manhua"
+			},
+			{
+				"type": "genre",
+				"name": "Manhwa",
+				"id": "Manhwa"
+			},
+			{
+				"type": "genre",
+				"name": "Martial Arts",
+				"id": "martial-arts"
+			},
+			{
+				"type": "genre",
+				"name": "Mecha",
+				"id": "mecha"
+			},
+			{
+				"type": "genre",
+				"name": "Medical",
+				"id": "medical"
+			},
+			{
+				"type": "genre",
+				"name": "Mystery",
+				"id": "mystery"
+			},
+			{
+				"type": "genre",
+				"name": "Rebirth",
+				"id": "rebirth"
+			},
+			{
+				"type": "genre",
+				"name": "Reincarnation",
+				"id": "reincarnation"
+			},
+			{
+				"type": "genre",
+				"name": "Romance",
+				"id": "romance"
+			},
+			{
+				"type": "genre",
+				"name": "School",
+				"id": "school"
+			},
+			{
+				"type": "genre",
+				"name": "School Life",
+				"id": "school-life"
+			},
+			{
+				"type": "genre",
+				"name": "Sci-Fi",
+				"id": "sci-fi"
+			},
+			{
+				"type": "genre",
+				"name": "Shoujo",
+				"id": "shoujo"
+			},
+			{
+				"type": "genre",
+				"name": "Shounen",
+				"id": "shounen"
+			},
+			{
+				"type": "genre",
+				"name": "Slice of Life",
+				"id": "slice-of-life"
+			},
+			{
+				"type": "genre",
+				"name": "Sports",
+				"id": "sports"
+			},
+			{
+				"type": "genre",
+				"name": "Super Power",
+				"id": "super-power"
+			},
+			{
+				"type": "genre",
+				"name": "Supernatural",
+				"id": "supernatural"
+			},
+			{
+				"type": "genre",
+				"name": "System",
+				"id": "system"
+			}
+		]
+	}
+]
