+++ src/rust/madara/sources/pojokmanga/res/settings.json
@@ -1,0 +1,16 @@
+[
+	{
+		"type": "group",
+		"title": "Settings",
+		"footer": "Note: Most Madara sources have a high resolution cover image and a low resolution cover image.\nEnabling High resolution cover image will increase the amount of data that is downloaded and may result in an increased loading time.",
+		"items": [
+			{
+				"type": "switch",
+				"key": "highres",
+				"title": "High Resolution Covers",
+				"subtitle": "For Home page and Search page",
+				"default": true
+			}
+		]
+	}
+]
