--- src/rust/mangastream/sources/manhwaland/res/source.json
@@ -1,21 +1,0 @@
-{
-	"info": {
-		"id": "id.manhwaland",
-		"lang": "id",
-		"name": "Manhwaland",
-		"version": 2,
-		"url": "https://manhwaland.mom",
-		"nsfw": 2
-	},
-	"listings": [
-		{
-			"name": "Popular"
-		},
-		{
-			"name": "Latest"
-		},
-		{
-			"name": "New"
-		}
-	]
-}
