+++ src/rust/mangastream/sources/manhwalist/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "manhwaland"
+name = "manhwalist"
 version = "0.1.0"
 edition = "2021"
 publish = false
